requirements.txt
pandas
numpy
tqdm
python-chess
faiss-cpu
torch
transformers
sentence-transformers
pillow
pandas
faiss-cpu
cairosvg
pyarrow